+++ tests/http/test_server.py
@@ -1,4 +1,7 @@
 import os
+import urllib
+import tempfile
+import shutil
 from unittest import mock
 
 import tornado.testing
@@ -17,6 +20,7 @@ class HttpServerTest(tornado.testing.AsyncHTTPTestCase):
                 "zeroconf": "",
                 "allowed_origins": [],
                 "csrf_protection": True,
+                "default_app": "mopidy",
             }
         }
 
@@ -59,7 +63,7 @@ class RootRedirectTest(HttpServerTest):
 class MopidyAppTest(HttpServerTest):
     def test_should_return_index(self):
         response = self.fetch("/mopidy/", method="GET")
-        body = tornado.escape.to_unicode(response.body)
+        body = response.body.decode()
 
         assert "This web server is a part of the Mopidy music server." in body
         assert "testapp" in body
@@ -76,7 +80,7 @@ class MopidyAppTest(HttpServerTest):
     def test_should_return_static_files(self):
         response = self.fetch("/mopidy/mopidy.css", method="GET")
 
-        assert "html {" in tornado.escape.to_unicode(response.body)
+        assert "html {" in response.body.decode()
         assert response.headers["X-Mopidy-Version"] == mopidy.__version__
         assert response.headers["Cache-Control"] == "no-cache"
 
@@ -85,18 +89,12 @@ class MopidyWebSocketHandlerTest(HttpServerTest):
     def test_should_return_ws(self):
         response = self.fetch("/mopidy/ws", method="GET")
 
-        assert (
-            'Can "Upgrade" only to "WebSocket".'
-            == tornado.escape.to_unicode(response.body)
-        )
+        assert 'Can "Upgrade" only to "WebSocket".' == response.body.decode()
 
     def test_should_return_ws_old(self):
         response = self.fetch("/mopidy/ws/", method="GET")
 
-        assert (
-            'Can "Upgrade" only to "WebSocket".'
-            == tornado.escape.to_unicode(response.body)
-        )
+        assert 'Can "Upgrade" only to "WebSocket".' == response.body.decode()
 
 
 class MopidyRPCHandlerTest(HttpServerTest):
@@ -268,7 +266,12 @@ class MopidyRPCHandlerNoCSRFProtectionTest(HttpServerTest):
 class HttpServerWithStaticFilesTest(tornado.testing.AsyncHTTPTestCase):
     def get_app(self):
         config = {
-            "http": {"hostname": "127.0.0.1", "port": 6680, "zeroconf": ""}
+            "http": {
+                "hostname": "127.0.0.1",
+                "port": 6680,
+                "zeroconf": "",
+                "default_app": "static",
+            }
         }
         core = mock.Mock()
 
@@ -313,7 +316,12 @@ def wsgi_app_factory(config, core):
 class HttpServerWithWsgiAppTest(tornado.testing.AsyncHTTPTestCase):
     def get_app(self):
         config = {
-            "http": {"hostname": "127.0.0.1", "port": 6680, "zeroconf": ""}
+            "http": {
+                "hostname": "127.0.0.1",
+                "port": 6680,
+                "zeroconf": "",
+                "default_app": "wsgi",
+            }
         }
         core = mock.Mock()
 
@@ -334,5 +342,200 @@ class HttpServerWithWsgiAppTest(tornado.testing.AsyncHTTPTestCase):
     def test_can_wrap_wsgi_apps(self):
         response = self.fetch("/wsgi/", method="GET")
 
+        assert response.code == 200
+        assert "Hello, world!" in response.body.decode()
+
+
+def default_webapp_factory(config, core):
+    class MainHandler(tornado.web.RequestHandler):
+        def get(self):
+            self.write("Hello from default webapp")
+
+    return [("/", MainHandler, {})]
+
+
+class HttpServerWithAppDefaultApp(tornado.testing.AsyncHTTPTestCase):
+    def get_app(self):
+        config = {
+            "http": {
+                "hostname": "127.0.0.1",
+                "port": 6680,
+                "zeroconf": "",
+                "default_app": "default_app",
+            }
+        }
+        core = mock.Mock()
+
+        apps = [dict(name="default_app", factory=default_webapp_factory)]
+
+        http_server = actor.HttpServer(
+            config=config, core=core, sockets=[], apps=apps, statics=[]
+        )
+
+        return tornado.web.Application(http_server._get_request_handlers())
+
+    def test_should_redirect_to_default_app(self):
+        response = self.fetch("/", method="GET", follow_redirects=False)
+
+        assert response.code == 302
+        assert response.headers["Location"] == "/default_app/"
+
+        response = self.fetch(
+            "/default_app/", method="GET", follow_redirects=True
+        )
+
+        assert response.code == 200
+        assert "Hello from default webapp" in response.body.decode()
+
+
+class HttpServerWithStaticDefaultApp(tornado.testing.AsyncHTTPTestCase):
+    def get_app(self):
+        config = {
+            "http": {
+                "hostname": "127.0.0.1",
+                "port": 6680,
+                "zeroconf": "",
+                "default_app": "default_app",
+            }
+        }
+        core = mock.Mock()
+
+        statics = [dict(name="default_app", path=os.path.dirname(__file__))]
+
+        http_server = actor.HttpServer(
+            config=config, core=core, sockets=[], apps=[], statics=statics
+        )
+
+        return tornado.web.Application(http_server._get_request_handlers())
+
+    def test_should_redirect_to_default_app(self):
+        response = self.fetch("/", method="GET", follow_redirects=False)
+
+        assert response.code == 302
+        assert response.headers["Location"] == "/default_app/"
+
+
+class HttpServerWithInvalidDefaultApp(HttpServerTest):
+    def get_config(self):
+        config = super(HttpServerWithInvalidDefaultApp, self).get_config()
+        config["http"]["default_app"] = "invalid_webclient"
+        return config
+
+    def test_should_redirect_to_clients_list(self):
+        response = self.fetch("/", method="GET", follow_redirects=False)
+
+        assert response.code == 302
+        assert response.headers["Location"] == "/mopidy/"
+
+        response = self.fetch("/", method="GET")
+        body = response.body.decode()
+
+        assert "This web server is a part of the Mopidy music server." in body
+
+        assert "testapp" in body
+        assert "teststatic" in body
+        assert response.headers["X-Mopidy-Version"] == mopidy.__version__
+        assert response.headers["Cache-Control"] == "no-cache"
+
+
+def cookie_secret_app_factory(config, core):
+    class BaseHandler(tornado.web.RequestHandler):
+        def get_current_user(self):
+            return self.get_secure_cookie("user")
+
+    class LoginHandler(BaseHandler):
+        def get(self):
+            self.write("This is a login form")
+
+        def post(self):
+            self.set_secure_cookie("user", self.get_argument("name"))
+            self.write("Logged in")
+
+    class MainHandler(BaseHandler):
+        def get(self):
+            if not self.current_user:
+                self.write("Unknown user...")
+                return
+
+            name = tornado.escape.xhtml_escape(self.current_user)
+            self.write("Hello, " + name)
+
+    return [("/", MainHandler, {}), ("/login", LoginHandler, {})]
+
+
+class HttpServerTestLoginWithSecureCookie(tornado.testing.AsyncHTTPTestCase):
+    def get_app(self):
+        self._dirpath = tempfile.mkdtemp()
+
+        config = {
+            "http": {
+                "hostname": "127.0.0.1",
+                "port": 6680,
+                "zeroconf": "",
+                "default_app": "mopidy",
+            },
+            "core": {"data_dir": self._dirpath},
+        }
+        core = mock.Mock()
+
+        apps = [{"name": "cookie_secret", "factory": cookie_secret_app_factory}]
+
+        http_server = actor.HttpServer(
+            config=config, core=core, sockets=[], apps=apps, statics=[]
+        )
+
+        return tornado.web.Application(
+            http_server._get_request_handlers(),
+            cookie_secret=http_server._get_cookie_secret(),
+        )
+
+    def test_main_access_without_login(self):
+        response = self.fetch("/cookie_secret", method="GET")
+
         assert 200 == response.code
-        assert "Hello, world!" in tornado.escape.to_unicode(response.body)
+        assert "Unknown user..." in response.body.decode()
+
+    def test_accessing_login_form_get(self):
+        response = self.fetch("/cookie_secret/login", method="GET")
+
+        assert 200 == response.code
+        assert "This is a login form" in response.body.decode()
+
+    def test_login(self):
+        post_data = {"name": "theuser"}
+        body = urllib.parse.urlencode(post_data)
+
+        response = self.fetch("/cookie_secret/login", method="POST", body=body)
+
+        assert 200 == response.code
+        assert "Logged in" in response.body.decode()
+
+        shutil.rmtree(self._dirpath)
+
+
+def test_get_secure_cookie(tmp_path):
+    config = {
+        "http": {
+            "hostname": "127.0.0.1",
+            "port": 6680,
+            "zeroconf": "",
+            "default_app": "mopidy",
+        },
+        "core": {"data_dir": tmp_path},
+    }
+    core = mock.Mock()
+
+    http_server = actor.HttpServer(
+        config=config, core=core, sockets=[], apps=[], statics=[]
+    )
+
+    # first secret, generating
+    secret_1 = http_server._get_cookie_secret()
+
+    assert isinstance(secret_1, str)
+    assert secret_1 != ""
+    assert len(secret_1) == 64
+
+    # second secret, from file
+    secret_2 = http_server._get_cookie_secret()
+    assert secret_1 == secret_2